requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.12
GitPython==3.1.44
GitPython==3.1.44
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.48
langchain_deeplake==0.1.0
langchain_openai==0.3.10
langchain_text_splitters==0.3.7
pydantic==2.10.6
python-dotenv==1.1.0
Requests==2.32.3
rich==13.9.4
schedule==1.2.2
uvicorn==0.34.0
